requirements.txt
pandas==2.2.2
scikit-learn==1.4.2
xgboost==3.0.2
optuna==4.4.0
mlflow==3.1.1
pytest==8.4.0
matplotlib==3.10.1
seaborn==0.13.2
kaggle==1.7.4.5
imblearn==0.0
PyYAML==6.0.1
pandera==0.20.0